Move PlayUpdateLight to chunk.py as it's for CHUNKS
